# isp-api-service/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: ISP Management System API
  version: 1.0.0
  description: |
    Kontrak API untuk Sistem Manajemen Terintegrasi ISP (MVP).
    Source of truth untuk komunikasi antara FE (portal pelanggan + dashboard admin)
    dan BE (isp-api-service :8080, isp-billing-service :8081).
    Endpoint pada path /billing/*, /internal/*, /webhook/*, /customers/status
    dilayani oleh isp-billing-service (:8081); sisanya oleh isp-api-service (:8080).
servers:
  - url: http://localhost:8080
    description: isp-api-service (dev)
  - url: http://localhost:8081
    description: isp-billing-service (dev)
...
paths:
...
  /billing/generate:
    post:
      tags: [invoices]
      summary: Generate tagihan massal (superadmin) — isp-billing-service (:8081)
      description: >
        Trigger manual. Cron otomatis berjalan tanggal 1 pukul 00:00.
        Idempoten per (customer, period); invoice yang sudah ada di-skip.
      security: [{ bearerAuth: [] }]
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                period: { type: string, description: "YYYY-MM (default: bulan berjalan)" }
      responses:
        '200':
          description: Ringkasan hasil generate
          content:
            application/json:
              schema:
                type: object
                required: [period, created, skipped]
                properties:
                  period: { type: string }
                  created: { type: integer }
                  skipped: { type: integer }
...
  /billing/customers/{id}/isolate:
    post:
      tags: [network]
      summary: Isolir manual — disable PPPoE (superadmin) — isp-billing-service (:8081)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Hasil isolir
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  isolated: { type: boolean }
  /billing/customers/{id}/reactivate:
    post:
      tags: [network]
      summary: Buka isolir manual — enable PPPoE (superadmin) — isp-billing-service (:8081)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Hasil buka isolir
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  reactivated: { type: boolean }
  /internal/customers/{id}/reactivate:
    post:
      tags: [network]
      summary: Buka isolir otomatis (dipanggil isp-api-service saat approve) — isp-billing-service (:8081)
      description: Diamankan header X-Internal-Secret (shared secret antar service), bukan JWT.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer, format: int64 }
      responses:
        '200':
          description: Hasil
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  reactivated: { type: boolean }
        '401': { description: Secret tidak valid }
  /webhook/network-status:
    post:
      tags: [network]
      summary: Webhook status koneksi dari Mikrotik — isp-billing-service (:8081)
      description: Diamankan header X-Webhook-Secret. Status alive→online, rto→offline.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                identifier: { type: string, description: "PPPoE / IP / MAC (atau pakai field spesifik di bawah)" }
                pppoe_username: { type: string }
                ip: { type: string }
                mac: { type: string }
                status: { type: string, enum: [alive, rto] }
      responses:
        '200':
          description: Status tercatat
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok: { type: boolean }
                  customer_id: { type: integer, format: int64 }
                  state: { type: string, enum: [online, offline] }
        '400': { description: Payload tidak valid }
        '401': { description: Secret tidak valid }
        '404': { $ref: '#/components/responses/NotFound' }
  /customers/status:
    get:
      tags: [network]
      summary: Snapshot status koneksi semua pelanggan (admin, polling 5s) — isp-billing-service (:8081)
      security: [{ bearerAuth: [] }]
      responses:
        '200':
          description: Map customerId → state. customer tanpa entri dianggap unknown.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: { type: string, enum: [online, offline] }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
...
    NotFound:
      description: Tidak ditemukan
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
...
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
        message: { type: string }